Separate network error and rethrow error from module error
While refactoring for move-away-from-Promise.all, I've started to notice
that having a `ModuleError` enum with `network` variant and `raw
exception` variant doesn't make sence because they actually have
different meanings.

So, in this commit, the `error` property will be separated into
`network_error` and `rethrow_error` so that we can check them
differently and no need to implement the confusing `PartialOrd` for
`ModuleError`.

Also, due to changing to separated errors, we will need to check if the
descendant or the module itself has `network_error` or not when the
module is trying to find the first parse error._